+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/src/main/resources/META-INF/resources/metal/js/components/Layout/index.es.js
@@ -1,5 +1,0 @@
-import LayoutRenderer from './LayoutRenderer.es';
-import LayoutSupport from './LayoutSupport.es';
-
-export default LayoutRenderer;
-export {LayoutSupport, LayoutRenderer};