+++ modules/apps/journal/journal-api/src/main/java/com/liferay/journal/service/JournalArticleServiceWrapper.java
@@ -650,8 +650,8 @@ public class JournalArticleServiceWrapper implements JournalArticleService,
 	* @param groupId the primary key of the web content article's group
 	* @param folderId the primary key of the web content article folder
 	* @return the matching web content articles
-	* @deprecated As of Judson (7.1.x), replaced by {@link
-	#getArticles(long groupId, long folderId, Locale locale)}
+	* @deprecated As of Judson (7.1.x), replaced by {@link #getArticles(long
+	groupId, long folderId, Locale locale)}
 	*/
 	@Deprecated
 	@Override
@@ -682,9 +682,9 @@ public class JournalArticleServiceWrapper implements JournalArticleService,
 	return (not inclusive)
 	* @param obc the comparator to order the web content articles
 	* @return the matching web content articles
-	* @deprecated As of Judson (7.1.x), replaced by {@link
-	#getArticles(long groupId, long folderId, Locale locale,
-	int start, int end, OrderByComparator obc)}
+	* @deprecated As of Judson (7.1.x), replaced by {@link #getArticles(long
+	groupId, long folderId, Locale locale, int start, int end,
+	OrderByComparator obc)}
 	*/
 	@Deprecated
 	@Override
@@ -1222,6 +1222,19 @@ public class JournalArticleServiceWrapper implements JournalArticleService,
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _journalArticleService.getLatestArticle(groupId, className,
 			classPK);
+	}
+
+	@Override
+	public java.util.List<com.liferay.journal.model.JournalArticle> getLatestArticles(
+		long groupId, int status, int start, int end,
+		com.liferay.portal.kernel.util.OrderByComparator<com.liferay.journal.model.JournalArticle> obc) {
+		return _journalArticleService.getLatestArticles(groupId, status, start,
+			end, obc);
+	}
+
+	@Override
+	public int getLatestArticlesCount(long groupId, int status) {
+		return _journalArticleService.getLatestArticlesCount(groupId, status);
 	}
 
 	@Override