+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/internal/exportimport/data/handler/JournalArticleStagedModelDataHandler.java
@@ -961,9 +961,6 @@ public class JournalArticleStagedModelDataHandler
 
 				_importAssetDisplayPage(
 					portletDataContext, article, importedArticle);
-
-				_importFriendlyURLEntries(
-					portletDataContext, article, importedArticle);
 			}
 			finally {
 				ServiceContextThreadLocal.popServiceContext();
@@ -1108,11 +1105,6 @@ public class JournalArticleStagedModelDataHandler
 		return fetchExistingArticle(
 			articleUuid, articleResourceUuid, companyGroup.getGroupId(),
 			articleId, newArticleId, version, preloaded);
-	}
-
-	@Override
-	protected String[] getSkipImportReferenceStagedModelNames() {
-		return _SKIP_IMPORT_REFERENCE_STAGED_MODEL_NAMES;
 	}
 
 	protected boolean isExpireAllArticleVersions(long companyId)
@@ -1270,8 +1262,11 @@ public class JournalArticleStagedModelDataHandler
 				article.getResourcePrimKey());
 
 		for (FriendlyURLEntry friendlyURLEntry : friendlyURLEntries) {
+			StagedModelDataHandlerUtil.exportStagedModel(
+				portletDataContext, friendlyURLEntry);
+
 			StagedModelDataHandlerUtil.exportReferenceStagedModel(
-				portletDataContext, article, friendlyURLEntry,
+				portletDataContext, friendlyURLEntry, article,
 				PortletDataContext.REFERENCE_TYPE_DEPENDENCY);
 		}
 	}
@@ -1316,37 +1311,6 @@ public class JournalArticleStagedModelDataHandler
 		}
 	}
 
-	private void _importFriendlyURLEntries(
-			PortletDataContext portletDataContext, JournalArticle article,
-			JournalArticle importedArticle)
-		throws PortalException {
-
-		List<Element> friendlyURLEntryElements =
-			portletDataContext.getReferenceDataElements(
-				article, FriendlyURLEntry.class);
-
-		for (Element friendlyURLEntryElement : friendlyURLEntryElements) {
-			String path = friendlyURLEntryElement.attributeValue("path");
-
-			FriendlyURLEntry friendlyURLEntry =
-				(FriendlyURLEntry)portletDataContext.getZipEntryAsObject(path);
-
-			friendlyURLEntry.setClassNameId(
-				_portal.getClassNameId(JournalArticle.class));
-			friendlyURLEntry.setClassPK(importedArticle.getResourcePrimKey());
-
-			StagedModelDataHandlerUtil.importStagedModel(
-				portletDataContext, friendlyURLEntry);
-		}
-
-		FriendlyURLEntry mainFriendlyURLEntry =
-			_friendlyURLEntryLocalService.getMainFriendlyURLEntry(
-				JournalArticle.class, importedArticle.getResourcePrimKey());
-
-		_journalArticleLocalService.updateArticle(
-			importedArticle.getId(), mainFriendlyURLEntry.getUrlTitle());
-	}
-
 	/**
 	 * @deprecated As of Judson (7.1.x), only used for backwards compatibility
 	 *             with LARs that use journal schema under 1.1.0
@@ -1368,9 +1332,6 @@ public class JournalArticleStagedModelDataHandler
 					article.getLegacyDescription()));
 		}
 	}
-
-	private static final String[] _SKIP_IMPORT_REFERENCE_STAGED_MODEL_NAMES =
-		{FriendlyURLEntry.class.getName()};
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		JournalArticleStagedModelDataHandler.class);