+++ modules/sdk/gradle-plugins/CHANGELOG.markdown
@@ -5410,10 +5410,42 @@ version 2.2.2.
 ## 3.12.110 - 2018-08-15
 
 ### Changed
-- [LPS-83790]: Update the default value of the
-`liferay.appServers.tomcat.version` property to `9.0.10`.
+- [LPS-83790]: Update the `liferay.appServers.tomcat.version` property's default
+value to `9.0.10`.
 - [LPS-83790]: Update the [Liferay Gradle Plugins Test Integration] dependency
 to version 2.3.1.
+
+## 3.12.111 - 2018-08-21
+
+### Changed
+- [LPS-84615]: Update the [Liferay Gradle Plugins Service Builder] dependency to
+version 2.1.55.
+- [LPS-84615]: Update the [Liferay Portal Tools Service Builder] dependency to
+version 1.0.238.
+
+## 3.12.112 - 2018-08-22
+
+### Changed
+- [LPS-84119]: Update the [Liferay Gradle Plugins Source Formatter] dependency
+to version 2.3.238.
+- [LPS-84119]: Update the [Liferay Source Formatter] dependency to version
+1.0.712.
+- [LPS-84218]: Update the [Liferay CSS Builder] dependency to version 2.1.3.
+- [LPS-84218]: Update the [Liferay Gradle Plugins CSS Builder] dependency to
+version 2.2.3.
+
+## 3.12.113 - 2018-08-23
+
+### Changed
+- [LPS-84119]: Update the [Liferay Gradle Plugins Source Formatter] dependency
+to version 2.3.239.
+- [LPS-84119]: Update the [Liferay Source Formatter] dependency to version
+1.0.713.
+
+## 3.12.114 - 2018-08-23
+
+### Changed
+- [LPS-83067]: Update the [Liferay Ant BND] dependency to version 2.0.54.
 
 [Gradle Bundle Plugin]: https://github.com/TomDmitriev/gradle-bundle-plugin
 [Liferay Alloy Taglib]: https://github.com/liferay/alloy-taglibs
@@ -5845,6 +5877,8 @@ to version 2.3.1.
 [LPS-84039]: https://issues.liferay.com/browse/LPS-84039
 [LPS-84119]: https://issues.liferay.com/browse/LPS-84119
 [LPS-84213]: https://issues.liferay.com/browse/LPS-84213
+[LPS-84218]: https://issues.liferay.com/browse/LPS-84218
 [LPS-84307]: https://issues.liferay.com/browse/LPS-84307
 [LPS-84473]: https://issues.liferay.com/browse/LPS-84473
+[LPS-84615]: https://issues.liferay.com/browse/LPS-84615
 [LRDOCS-3023]: https://issues.liferay.com/browse/LRDOCS-3023